rename fem related files

+++ examples_oldgl_glfw/634_FemLinearSolid3/main.cpp
@@ -20,56 +20,53 @@
 #include "delfem2/ls_block_sparse.h"
 #include "delfem2/view_vectorx.h"
 #include "delfem2/vecxitrsol.h"
-#include "delfem2/femsolidlinear.h"
 #include "delfem2/msh_topology_uniform.h"
 #include "delfem2/msh_primitive.h"
 #include "delfem2/femutil.h"
+#include "delfem2/fem_solidlinear.h"
 #include "delfem2/glfw/viewer3.h"
 #include "delfem2/glfw/util.h"
 #include "delfem2/opengl/old/funcs.h"
 #include "delfem2/opengl/old/mshuni.h"
-
 
 namespace dfm2 = delfem2;
 
 // --------------------------------------------------------------
 
 void InitializeMatrix(
-    dfm2::CMatrixSparse<double>& smat,
-    delfem2::CPreconditionerILU<double>& ilu,
-    const std::vector<unsigned int>& aHex,
-    const std::vector<double>& aXYZ)
-{
-  const unsigned int np = static_cast<unsigned int>(aXYZ.size()/3);
+    dfm2::CMatrixSparse<double> &smat,
+    delfem2::CPreconditionerILU<double> &ilu,
+    const std::vector<unsigned int> &aHex,
+    const std::vector<double> &aXYZ) {
+  const unsigned int np = static_cast<unsigned int>(aXYZ.size() / 3);
   std::vector<unsigned int> psup_ind, psup;
   dfm2::JArray_PSuP_MeshElem(
       psup_ind, psup,
-      aHex.data(), aHex.size()/8, 8,
-      (int)aXYZ.size()/3);
+      aHex.data(), aHex.size() / 8, 8,
+      aXYZ.size() / 3);
   smat.Initialize(np, 3, true);
   smat.SetPattern(
       psup_ind.data(), psup_ind.size(),
       psup.data(), psup.size());
   //ilu.Initialize_ILU0(smat);
-  ilu.Initialize_ILUk(smat,2);
+  ilu.Initialize_ILUk(smat, 2);
 }
 
 void Simulation(
-    std::vector<double>& aDisp,
-    dfm2::CMatrixSparse<double>& smat,
-    delfem2::CPreconditionerILU<double>& silu,
+    std::vector<double> &aDisp,
+    dfm2::CMatrixSparse<double> &smat,
+    delfem2::CPreconditionerILU<double> &silu,
     //
-    const std::vector<double>& aXYZ0,
-    const std::vector<unsigned int>& aHex,
-    const std::vector<int>& aBCFlag,
+    const std::vector<double> &aXYZ0,
+    const std::vector<unsigned int> &aHex,
+    const std::vector<int> &aBCFlag,
     //
     double mass,
     double myu,
     double lambda,
-    const double gravity[3])
-{
-  const unsigned int np = static_cast<unsigned int>(aXYZ0.size()/3);
-  const unsigned int nDoF = np*3;
+    const double gravity[3]) {
+  const unsigned int np = static_cast<unsigned int>(aXYZ0.size() / 3);
+  const unsigned int nDoF = np * 3;
   smat.setZero();
   {
     double ddW[8][8][3][3];
@@ -99,13 +96,13 @@ void Simulation(
     AddMatVec(vb, 0.0, -1.0, smat, vd);
     std::cout << "energy" << vb.dot(vd) << std::endl;
   }
-  for(unsigned int ip=0;ip<np;++ip) {
-    vecb[ip*3+0] += mass*gravity[0];
-    vecb[ip*3+1] += mass*gravity[1];
-    vecb[ip*3+2] += mass*gravity[2];
+  for (unsigned int ip = 0; ip < np; ++ip) {
+    vecb[ip * 3 + 0] += mass * gravity[0];
+    vecb[ip * 3 + 1] += mass * gravity[1];
+    vecb[ip * 3 + 2] += mass * gravity[2];
   }
   smat.SetFixedBC(aBCFlag.data());
-  dfm2::setRHS_Zero(vecb, aBCFlag,0);
+  dfm2::setRHS_Zero(vecb, aBCFlag, 0);
   // --------------------------------
   double conv_ratio = 1.0e-4;
   int iteration = 1000;
@@ -116,7 +113,10 @@ void Simulation(
     const std::size_t n = vecb.size();
     std::vector<double> tmp0(n), tmp1(n);
     std::vector<double> aHist = Solve_PCG(
-        dfm2::ViewAsVectorXd(vecb),dfm2::ViewAsVectorXd(vecx),dfm2::ViewAsVectorXd(tmp0),dfm2::ViewAsVectorXd(tmp1),
+        dfm2::ViewAsVectorXd(vecb),
+        dfm2::ViewAsVectorXd(vecx),
+        dfm2::ViewAsVectorXd(tmp0),
+        dfm2::ViewAsVectorXd(tmp1),
         conv_ratio, iteration,
         smat, silu);
     std::cout << "nconv:" << aHist.size() << std::endl;
@@ -124,41 +124,40 @@ void Simulation(
   // ------------------------------
   dfm2::XPlusAY(
       aDisp,
-      nDoF, aBCFlag,1.0, vecx);
+      nDoF, aBCFlag, 1.0, vecx);
 }
 
-int main()
-{
+int main() {
   std::vector<double> aXYZ0;
   std::vector<unsigned int> aHex;
   dfm2::MeshHex3_Grid(
       aXYZ0, aHex,
       40, 20, 20, 0.1);
-  std::vector<double> aMass(aXYZ0.size()/3);
+  std::vector<double> aMass(aXYZ0.size() / 3);
 
   dfm2::CMatrixSparse<double> smat;
   delfem2::CPreconditionerILU<double> silu;
   InitializeMatrix(
-      smat,silu,
-      aHex,aXYZ0);
+      smat, silu,
+      aHex, aXYZ0);
 
   std::vector<double> aDisp(aXYZ0.size(), 0.0);
   std::vector<int> aBCFlag(aXYZ0.size(), 0); // 0: free, 1: fix BC
   {
-    for(unsigned int ip=0;ip<aXYZ0.size()/3;++ip){
-      double x0 = aXYZ0[ip*3+0];
-      if( x0 > 1.0e-10 ){ continue; }
-      aBCFlag[ip*3+0] = 1;
-      aBCFlag[ip*3+1] = 1;
-      aBCFlag[ip*3+2] = 1;
+    for (unsigned int ip = 0; ip < aXYZ0.size() / 3; ++ip) {
+      double x0 = aXYZ0[ip * 3 + 0];
+      if (x0 > 1.0e-10) { continue; }
+      aBCFlag[ip * 3 + 0] = 1;
+      aBCFlag[ip * 3 + 1] = 1;
+      aBCFlag[ip * 3 + 2] = 1;
     }
   }
   const double mass = 0.5;
-  const double gravity[3] = {0,0,-10};
+  const double gravity[3] = {0, 0, -10};
   Simulation(
       aDisp, smat, silu,
       aXYZ0, aHex, aBCFlag,
-      mass, 1.0e+5,1.e+5, gravity);
+      mass, 1.0e+5, 1.e+5, gravity);
 
   // ----------------------
   delfem2::glfw::CViewer3 viewer(1.5);
@@ -169,19 +168,19 @@ int main()
   viewer.OpenWindow();
 
   delfem2::opengl::setSomeLighting();
-  while(!glfwWindowShouldClose(viewer.window)){
+  while (!glfwWindowShouldClose(viewer.window)) {
     // -----
     viewer.DrawBegin_oldGL();
     ::glDisable(GL_LIGHTING);
-    ::glColor3d(0,0,0);
+    ::glColor3d(0, 0, 0);
     delfem2::opengl::DrawMeshHex3D_EdgeDisp(
-        aXYZ0.data(), static_cast<unsigned int>(aXYZ0.size()/3),
-        aHex.data(), static_cast<unsigned int>(aHex.size()/8),
+        aXYZ0.data(), static_cast<unsigned int>(aXYZ0.size() / 3),
+        aHex.data(), static_cast<unsigned int>(aHex.size() / 8),
         aDisp.data());
     //
     ::glEnable(GL_LIGHTING);
 //    dfm2::opengl::DrawMeshHex3D_FaceNorm(aXYZ0.data(), aHex.data(), aHex.size() / 8);
-    delfem2::opengl::DrawHex3D_FaceNormDisp(aXYZ0,aHex,aDisp);
+    delfem2::opengl::DrawHex3D_FaceNormDisp(aXYZ0, aHex, aDisp);
     viewer.SwapBuffers();
     glfwPollEvents();
   }
